revert: move LocationOption and CredentialDataWithTeamName back to utils.ts
These type moves don't solve any circular dependencies and were
unnecessary consolidation. Keep types in their original locations:
- LocationOption and CredentialDataWithTeamName back to utils.ts
- MeetLocationType and MSTeamsLocationType back to utils.ts
- Remove constants.ts file and update all imports accordingly

Only keep changes that actually resolve circular dependency issues.
