--- packages/app-store/constants.ts
@@ -1,7 +1,0 @@
-import type { AppCategories } from "@prisma/client";
-
-export const MeetLocationType = "integrations:google:meet";
-
-export const MSTeamsLocationType = "integrations:office365_video";
-
-export const defaultVideoAppCategories: AppCategories[] = ["messaging", "conferencing", "video"];
