+++ packages/app-store/types.d.ts
@@ -22,30 +22,6 @@ export type CredentialOwner = {
   readOnly?: boolean;
 };
 
-export type LocationOption = {
-  label: string;
-  value: string;
-  icon?: string;
-  disabled?: boolean;
-};
-
-export type CredentialDataWithTeamName = {
-  id: number;
-  type: string;
-  key: any;
-  userId: number | null;
-  user: { email: string } | null;
-  teamId: number | null;
-  appId: string | null;
-  invalid: boolean | null;
-  delegatedTo: any;
-  delegatedToId?: string | null | undefined;
-  delegationCredentialId: string | null;
-  team?: {
-    name: string;
-  } | null;
-};
-
 export type EventTypeAppCardApp = RouterOutputs["viewer"]["apps"]["integrations"]["items"][number] & {
   credentialOwner?: CredentialOwner;
   credentialIds?: number[];
